Move unit test suites into `$(PROJROOT)/Tests/UnitTests`
Relates to [MACCAN-111]
