+++ Tests/UnitTests/Testing.mm
@@ -2,7 +2,7 @@
 //
 //  CAN Interface API, Version 3 (Testing)
 //
-//  Copyright (c) 2004-2021 Uwe Vogt, UV Software, Berlin (info@uv-software.com)
+//  Copyright (c) 2004-2022 Uwe Vogt, UV Software, Berlin (info@uv-software.com)
 //  All rights reserved.
 //
 //  This file is part of CAN API V3.
@@ -49,11 +49,11 @@
 #import "can_api.h"
 #import <XCTest/XCTest.h>
 
-@interface test_can_version : XCTestCase
+@interface Testing : XCTestCase
 
 @end
 
-@implementation test_can_version
+@implementation Testing
 
 - (void)setUp {
     // Put setup code here. This method is called before the invocation of each test method in the class.
@@ -64,27 +64,16 @@
     (void)can_exit(CANKILL_ALL);
 }
 
-// @xctest TC15.1: Query CAN API version number at any place in a standard sequence (cf. SmokeTest).
-//
-// @expected: Not NULL
-//
-- (void)testAnyPlaceAnyTime {
+- (void)testSmokeTest {
     can_bitrate_t bitrate = { TEST_BTRINDEX };
     can_status_t status = { CANSTAT_RESET };
-    char *string = NULL;
     int handle = INVALID_HANDLE;
     int rc = CANERR_FATAL;
-
+    
     // @test:
-    // @- get CAN API version of DUT1
-    string = can_version();
-    XCTAssertTrue(NULL != string);
     // @- initialize DUT1 with configured settings
     handle = can_init(DUT1, TEST_CANMODE, NULL);
     XCTAssertLessThanOrEqual(0, handle);
-    // @- get CAN API version of DUT1
-    string = can_version();
-    XCTAssertTrue(NULL != string);
     // @- get status of DUT1 and check to be in INIT state
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
@@ -92,29 +81,22 @@
     // @- start DUT1 with configured bit-rate settings
     rc = can_start(handle, &bitrate);
     XCTAssertEqual(CANERR_NOERROR, rc);
-    // @- get CAN API version of DUT1
-    string = can_version();
-    XCTAssertTrue(NULL != string);
     // @- get status of DUT1 and check to be in RUNNING state
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
     XCTAssertFalse(status.can_stopped);
-    // @- sunnyday traffic (optional):
-#if (SEND_TEST_FRAMES != 0)
+    // @- send some frames to DUT2 and receive some frames from DUT2
     CTester tester;
-    XCTAssertEqual(TEST_FRAMES, tester.SendSomeFrames(handle, DUT2, TEST_FRAMES));
-    XCTAssertEqual(TEST_FRAMES, tester.ReceiveSomeFrames(handle, DUT2, TEST_FRAMES));
+    NSLog(@"Be patient...");
+    XCTAssertEqual(TEST_TRAFFIC, tester.SendSomeFrames(handle, DUT2, TEST_TRAFFIC));
+    XCTAssertEqual(TEST_TRAFFIC, tester.ReceiveSomeFrames(handle, DUT2, TEST_TRAFFIC));
     // @- get status of DUT1 and check to be in RUNNING state
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
     XCTAssertFalse(status.can_stopped);
-#endif
     // @- stop/reset DUT1
     rc = can_reset(handle);
     XCTAssertEqual(CANERR_NOERROR, rc);
-    // @- get CAN API version of DUT1
-    string = can_version();
-    XCTAssertTrue(NULL != string);
     // @- get status of DUT1 and check to be in INIT state
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
@@ -122,11 +104,21 @@
     // @- shutdown DUT1
     rc = can_exit(handle);
     XCTAssertEqual(CANERR_NOERROR, rc);
-    // @- get CAN API version of DUT1
-    string = can_version();
-    XCTAssertTrue(NULL != string);
 }
+
+//- (void)testExample {
+//    // This is an example of a functional test case.
+//    // Use XCTAssert and related functions to verify your tests produce the correct results.
+//    XCTAssertTrue(true);
+//}
+
+//- (void)testPerformanceExample {
+//    // This is an example of a performance test case.
+//    [self measureBlock:^{
+//        // Put the code you want to measure the time of here.
+//    }];
+//}
 
 @end
 
-// $Id: test_can_version.mm 1035 2021-12-21 12:03:27Z makemake $  Copyright (c) UV Software, Berlin //
+// $Id: Testing.mm 1075 2022-01-04 22:00:43Z makemake $  Copyright (c) UV Software, Berlin //
