+++ Tests/UnitTests/ToDo.txt
@@ -1,0 +1,95 @@
+TC02.8: Check if interface can be initialized with operation mode bit MON set (listen-only mode):
+- try to send a frame & receive some frames
+TC02.9: Check if interface can be initialized with operation mode bit ERR set (error frame reception):
+- receive some error frames
+TC02.10: Check if interface can be initialized with operation mode bit NRTR set (suppress remote frames):
+- try to request & receive some remote frames
+TC02.11: Check if interface can be initialized with operation mode bit NXTD set (suppress extended frames):
+- try to send & receive some extended frames
+TC02.12: Check if interface can be initialized with operation mode bit FDOE set (CAN FD operation enabled):
+- try to send & receive some CAN FD long frames
+TC02.13: Check if interface can be initialized with operation mode bit FDOE and BRSE set (CAN FD operation with bit-rate switching enabled):
+- try to send & receive some CAN FD fast frames
+TC03.19: testWithValidCan20BitrateSettings
+- define test subject
+- implement this
+TC03.20: testWithInvalidCan20BitrateSettings
+- define test subject
+- implement this
+TC03.21: testWithSameCan20BitrateSettingsAfterCanStopped
+- define test subject
+- implement this
+TC03.22: testWithDifferentCan20BitrateSettingsAfterCanStopped
+- define test subject
+- implement this
+TC03.23: testWithValidCanFdBitrateSettings
+- define test subject
+- implement this
+TC03.24: testWithInvalidCanFdBitrateSettings
+- define test subject
+- implement this
+TC03.25: testWithSameCanFdBitrateSettingsAfterCanStopped
+- define test subject
+- implement this
+TC03.26: testWithDifferentCanFdBitrateSettingsAfterCanStopped
+- define test subject
+- implement this
+TC03.27: testWithCan20CiaIndexInCanFdMode
+- define test subject
+- implement this
+TC03.28: testWithCan20BitrateSettingsInCanFdMode
+- define test subject
+- implement this
+TC03.29: testWithCanFdBitrateSettingsInCan20Mode
+- define test subject
+- implement this
+TC04.9: testWhenMessageLost
+- define test subject
+- open point: How to loose a message?
+TC04.10: testTimestampAccuracy
+- define test case
+- implement this
+TC09.8: testWhenStatusBusOff
+- define test subject
+- implement this
+TC09.9: testWhenStatusWarningLevel
+- define test subject
+- implement this
+TC09.10: testWhenStatusBusBrror
+- define test subject
+- implement this
+TC09.11: testWhenStatusTransmitterBusy
+- define test subject
+- implement this
+TC09.12: testWhenStatusReceiverEmpty
+- define test subject
+- implement this
+TC09.13: testWhenStatusMessageLost
+- define test subject
+- implement this
+TC09.14: testWhenStatusQueueOverrun
+- define test subject
+- implement this
+TC10.10: testWhenStatusBusOff
+- define test subject
+- implement this
+TC10.11: testWhenStatusWarningLevel
+- define test subject
+- implement this
+TC10.12: testWhenStatusBusBrror
+- define test subject
+- implement this
+TC10.13: testWhenStatusTransmitterBusy
+- define test subject
+- implement this
+TC10.14: testWhenStatusReceiverEmpty
+- define test subject
+- implement this
+TC10.15: testWhenStatusMessageLost
+- define test subject
+- implement this
+TC10.16: testWhenStatusQueueOverrun
+- define test subject
+- implement this
+TC11.11: Get CAN bit-rate settings when CAN controller started with various CAN FD bit-rate settings and check for correctness:
+- implement this
