+++ Tests/UnitTests/test_can_firmware.mm
@@ -2,7 +2,7 @@
 //
 //  CAN Interface API, Version 3 (Testing)
 //
-//  Copyright (c) 2004-2021 Uwe Vogt, UV Software, Berlin (info@uv-software.com)
+//  Copyright (c) 2004-2022 Uwe Vogt, UV Software, Berlin (info@uv-software.com)
 //  All rights reserved.
 //
 //  This file is part of CAN API V3.
@@ -49,11 +49,11 @@
 #import "can_api.h"
 #import <XCTest/XCTest.h>
 
-@interface test_can_hardware : XCTestCase
+@interface test_can_firmware : XCTestCase
 
 @end
 
-@implementation test_can_hardware
+@implementation test_can_firmware
 
 - (void)setUp {
     // Put setup code here. This method is called before the invocation of each test method in the class.
@@ -64,7 +64,7 @@
     (void)can_exit(CANKILL_ALL);
 }
 
-// @xctest TC13.1: Get hardware version with invalid interface handle(s).
+// @xctest TC14.1: Get firmware version with invalid interface handle(s).
 //
 // @expected: NULL
 //
@@ -75,12 +75,14 @@
     int handle = INVALID_HANDLE;
     int rc = CANERR_FATAL;
 
-    // @test:
+    // @pre:
     // @- initialize DUT1 with configured settings
     handle = can_init(DUT1, TEST_CANMODE, NULL);
     XCTAssertLessThanOrEqual(0, handle);
+
+    // @test:
     // @- try to get version of DUT1 with invalid handle -1
-    string = can_hardware(INVALID_HANDLE);
+    string = can_firmware(INVALID_HANDLE);
     XCTAssertTrue(NULL == string);
     // @- get status of DUT1 and check to be in INIT state
     rc = can_status(handle, &status.byte);
@@ -90,13 +92,13 @@
     rc = can_start(handle, &bitrate);
     XCTAssertEqual(CANERR_NOERROR, rc);
     // @- try to get version of DUT1 with invalid handle INT32_MIN
-    string = can_hardware(INT32_MAX);
+    string = can_firmware(INT32_MIN);
     XCTAssertTrue(NULL == string);
     // @- get status of DUT1 and check to be in RUNNING state
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
     XCTAssertFalse(status.can_stopped);
-    // @- sunnyday traffic (optional):
+    // @- send and receive some frames to/from DUT2 (optional)
 #if (SEND_TEST_FRAMES != 0)
     CTester tester;
     XCTAssertEqual(TEST_FRAMES, tester.SendSomeFrames(handle, DUT2, TEST_FRAMES));
@@ -109,19 +111,21 @@
     // @- stop/reset DUT1
     rc = can_reset(handle);
     XCTAssertEqual(CANERR_NOERROR, rc);
-    // @- try to get version of DUT1 with invalid handle INT32_MIN
-    string = can_hardware(INT32_MIN);
+    // @- try to get version of DUT1 with invalid handle INT32_MAX
+    string = can_firmware(INT32_MAX);
     XCTAssertTrue(NULL == string);
     // @- get status of DUT1 and check to be in INIT state
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
     XCTAssertTrue(status.can_stopped);
+
+    // @post:
     // @- shutdown DUT1
     rc = can_exit(handle);
     XCTAssertEqual(CANERR_NOERROR, rc);
 }
 
-// @xctest TC13.2: Get hardware version when interface is not initialized.
+// @xctest TC14.2: Get firmware version when interface is not initialized.
 //
 // @expected: NULL
 //
@@ -134,15 +138,16 @@
 
     // @test:
     // @- try to get version of DUT1 with invalid handle -1
-    string = can_hardware(INVALID_HANDLE);
+    string = can_firmware(INVALID_HANDLE);
     XCTAssertTrue(NULL == string);
     // @- try to get version of DUT1 with invalid handle INT32_MIN
-    string = can_hardware(INT32_MAX);
+    string = can_firmware(INT32_MIN);
     XCTAssertTrue(NULL == string);
-    // @- try to get version of DUT1 with invalid handle INT32_MIN
-    string = can_hardware(INT32_MIN);
+    // @- try to get version of DUT1 with invalid handle INT32_MAX
+    string = can_firmware(INT32_MAX);
     XCTAssertTrue(NULL == string);
 
+    // @post:
     // @- initialize DUT1 with configured settings
     handle = can_init(DUT1, TEST_CANMODE, NULL);
     XCTAssertLessThanOrEqual(0, handle);
@@ -157,7 +162,7 @@
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
     XCTAssertFalse(status.can_stopped);
-    // @- sunnyday traffic (optional):
+    // @- send and receive some frames to/from DUT2 (optional)
 #if (SEND_TEST_FRAMES != 0)
     CTester tester;
     XCTAssertEqual(TEST_FRAMES, tester.SendSomeFrames(handle, DUT2, TEST_FRAMES));
@@ -179,7 +184,7 @@
     XCTAssertEqual(CANERR_NOERROR, rc);
 }
 
-// @xctest TC13.3: Get hardware version when interface already shutdown.
+// @xctest TC14.3: Get firmware version when interface already shutdown.
 //
 // @expected: NULL
 //
@@ -205,7 +210,7 @@
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
     XCTAssertFalse(status.can_stopped);
-    // @- sunnyday traffic (optional):
+    // @- send and receive some frames to/from DUT2 (optional)
 #if (SEND_TEST_FRAMES != 0)
     CTester tester;
     XCTAssertEqual(TEST_FRAMES, tester.SendSomeFrames(handle, DUT2, TEST_FRAMES));
@@ -228,11 +233,11 @@
 
     // @test:
     // @- try to get version of DUT1
-    string = can_hardware(handle);
+    string = can_firmware(handle);
     XCTAssertTrue(NULL == string);
 }
 
-// @xctest TC13.4: Query hardware version at any place in a standard sequence (cf. SmokeTest).
+// @xctest TC14.4: Query firmware version at any place in a standard sequence (cf. SmokeTest).
 //
 // @expected: Not NULL
 //
@@ -247,8 +252,8 @@
     // @- initialize DUT1 with configured settings
     handle = can_init(DUT1, TEST_CANMODE, NULL);
     XCTAssertLessThanOrEqual(0, handle);
-    // @- get hardware version of DUT1
-    string = can_hardware(handle);
+    // @- get firmware version of DUT1
+    string = can_firmware(handle);
     XCTAssertTrue(NULL != string);
     // @- get status of DUT1 and check to be in INIT state
     rc = can_status(handle, &status.byte);
@@ -257,14 +262,14 @@
     // @- start DUT1 with configured bit-rate settings
     rc = can_start(handle, &bitrate);
     XCTAssertEqual(CANERR_NOERROR, rc);
-    // @- get hardware version of DUT1
-    string = can_hardware(handle);
+    // @- get firmware version of DUT1
+    string = can_firmware(handle);
     XCTAssertTrue(NULL != string);
     // @- get status of DUT1 and check to be in RUNNING state
     rc = can_status(handle, &status.byte);
     XCTAssertEqual(CANERR_NOERROR, rc);
     XCTAssertFalse(status.can_stopped);
-    // @- sunnyday traffic (optional):
+    // @- send and receive some frames to/from DUT2 (optional)
 #if (SEND_TEST_FRAMES != 0)
     CTester tester;
     XCTAssertEqual(TEST_FRAMES, tester.SendSomeFrames(handle, DUT2, TEST_FRAMES));
@@ -277,8 +282,8 @@
     // @- stop/reset DUT1
     rc = can_reset(handle);
     XCTAssertEqual(CANERR_NOERROR, rc);
-    // @- get hardware version of DUT1
-    string = can_hardware(handle);
+    // @- get firmware version of DUT1
+    string = can_firmware(handle);
     XCTAssertTrue(NULL != string);
     // @- get status of DUT1 and check to be in INIT state
     rc = can_status(handle, &status.byte);
@@ -291,4 +296,4 @@
 
 @end
 
-// $Id: test_can_hardware.mm 1035 2021-12-21 12:03:27Z makemake $  Copyright (c) UV Software, Berlin //
+// $Id: test_can_firmware.mm 1086 2022-01-09 20:01:00Z haumea $  Copyright (c) UV Software, Berlin //
